Organização inicial create e update

+++ src/main/java/br/com/alelo/consumer/consumerpat/respository/ConsumerRepository.java
@@ -4,7 +4,7 @@ import br.com.alelo.consumer.consumerpat.entity.Consumer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ConsumerRepository extends JpaRepository<Consumer, Integer> {
+public interface ConsumerRepository extends JpaRepository<Consumer, Long> {
 
     @Query(nativeQuery = true, value = "select * from Consumer where FOOD_CARD_NUMBER = ? ")
     Consumer findByFoodCardNumber(int cardNumber);
